sam/requirements.txt
Django==5.0.1
google-generativeai==0.3.2
python-dotenv==1.0.0
django-cors-headers==4.3.1